feat: 예약 상세 및 결제 페이지 구현, 모달 컴포넌트 추가
- 예약 상세 페이지에서 결제 페이지로의 이동 기능 추가
- 결제 페이지에서 결제 수단 선택 및 결제 처리 기능 구현
- 예약 상세 페이지에 예약 취소 및 환불 모달 추가
- EditProfileModal, CancellationModal, RefundModal, ReservationActionModal, ReviewModal 컴포넌트 추가 및 통합
- 모달 컴포넌트들을 shared/ui/modal 디렉토리로 정리하여 관리 용이성 향상
- 예약 리스트 취소 버튼 삭제
