+++ apps/src/shared/ui/index.ts
@@ -1,1 +1,4 @@
 export {};
+
+// 모달 컴포넌트들
+export * from './modal';
